Issue Resolved of Controller call and Reformatted Project Structure Added Registration Page
