+++ Views/Home/Index.cshtml
@@ -3,7 +3,7 @@
 }
 
 <head>
-
+  <base href="/"/>
 </head>
 
 <body>
