+++ dotnetcoreandAngular5/AngularApp/src/app/Registration/Registration.component.ts
@@ -1,0 +1,10 @@
+import { Component, OnInit, Inject } from '@angular/core';
+
+@Component({
+  selector: 'registration',
+  templateUrl: './registration.html',
+})
+export class RegistrationComponent {
+
+
+}
