+++ dotnetcoreandAngular5/AngularApp/src/app/Registration/RegistrationPage/registrationPage.css
@@ -1,0 +1,2 @@
+﻿body {
+}
