+++ dotnetcoreandAngular5/AngularApp/src/app/Registration/RegistrationPage/registrationPage.ts
@@ -1,0 +1,10 @@
+import { Component, OnInit, Inject } from '@angular/core';
+
+@Component({
+  templateUrl: './registrationPage.html',
+})
+export class RegistrationPageComponent {
+
+  name: string;
+  age: number;
+}
