+++ dotnetcoreandAngular5/AngularApp/src/app/Registration/registration.css
@@ -1,0 +1,2 @@
+body {
+}
