+++ dotnetcoreandAngular5/AngularApp/src/app/kendo-component/componentList/kendo-component.css
@@ -1,0 +1,16 @@
+kendo-tabstrip p {
+  margin: 0;
+  padding: 8px;
+}
+
+kendo-splitter-pane {
+  padding: 16px;
+}
+
+button, input {
+  margin: 1em 0;
+}
+
+.k-gauge {
+  display: block;
+}
