+++ dotnetcoreandAngular5/AngularApp/src/app/kendo-component/kendo.component.ts
@@ -1,0 +1,8 @@
+import { Component, OnInit, Inject } from '@angular/core';
+
+@Component({
+  templateUrl: './kendo.html'
+})
+export class KendoComponent implements OnInit {
+  ngOnInit() { }
+}
