+++ dotnetcoreandAngular5/AngularApp/src/app/kendo-component/kendo.html
@@ -1,1 +1,2 @@
+
 <router-outlet></router-outlet>
